requirements.txt
numpy>=1.24.2
pandas>=2.0.0
matplotlib>=3.4
# seaborn>=0.11
python-dotenv>=1.0.0
openai>=1.2.4
pyyaml>=6.0
scikit-learn>=1.2.2
click>=8.0.1
urllib3>=1.26.17
jsonschema==4.20.0
injector>=0.21.0
ijson>=3.2.3
requests>=2.31.0

# Code Execution related
ipykernel==6.26.0
pre-commit>=2.19.0
tenacity>=8.2.2
plotly>=5.14.1
pytest>=7.0.0
vcrpy>=5.0.0
colorama>=0.4.6
# opencv-python
pdfplumber
pdf2image
pdfminer.six
pymupdf
pikepdf
ipywidgets
pypdf
Pillow
chainlit
openpyxl

## Use 2023-10-31-preview
azure-core
azure-ai-formrecognizer
azure-ai-documentintelligence
azure-search-documents
python-pptx
python-docx 
# comtypes
urllib3
json_repair
# pysqlite3 #required for chroma
chromadb
chainlit
mermaid-python
# pygraphviz
streamlit
accelerate
tiktoken
pyautogen
langchain
debugpy
#numexpr required for Langchain
numexpr 
langchain_openai
tabulate